begin to deploy

+++ app/(protected)/(tasks)/actions.ts
@@ -1,8 +1,9 @@
 "use server";
 
 import { prisma } from "@/lib/prisma";
+import { createServerSupabase } from "@/lib/supabase/server";
 import { TaskCreateSchema } from "@/lib/validations";
-import { revalidateTag } from "next/cache";
+import { revalidatePath, revalidateTag } from "next/cache";
 
 type ReturnState = {
   success: boolean;
@@ -13,6 +14,11 @@ export async function createTask(_prevState: ReturnState, formData: FormData) {
   const title = formData.get("title") as string;
   const description = formData.get("description") as string;
   const priority = formData.get("priority") as string;
+
+  const supabase = await createServerSupabase();
+
+  const { data } = await supabase.auth.getUser();
+  if (!data.user) throw new Error("Unauthorized");
 
   try {
     const saveData = TaskCreateSchema.parse({
@@ -26,16 +32,26 @@ export async function createTask(_prevState: ReturnState, formData: FormData) {
         title: saveData.title,
         description: saveData.description ?? "",
         priority: saveData.priority,
+        userId: data.user.id,
       },
     });
 
     revalidateTag("tasks", "max");
+    console.log("success to create" + saveData.title);
 
-    return { success: true, message: "Task created successfully" };
+    return {
+      success: true,
+      message: "Task created successfully",
+      version: Date.now(),
+    };
   } catch (error) {
-    console.error("Error creating task:", error);
+    console.log("Error creating task:", error);
     // Return readable error for client
-    return { success: false, message: "Failed to create task" };
+    return {
+      success: false,
+      message: "Failed to create task",
+      version: Date.now(),
+    };
   }
 }
 
@@ -67,11 +83,17 @@ export async function updateTitle(taskId: string, title: string) {
 
 export async function deleteTask(taskId: string) {
   try {
+    const supabase = await createServerSupabase();
+
+    const { data } = await supabase.auth.getUser();
+
+    if (!data.user) throw new Error("Unauthorized");
+
     await prisma.task.delete({
       where: { id: taskId },
     });
 
-    revalidateTag("tasks", "max");
+    revalidatePath("/tasks");
   } catch (error) {
     console.log(error);
   }
